refactor: separate fastn-spec-viewer into dedicated crate
- Create separate fastn-spec-viewer crate for clean architecture
- Move spec viewer functionality out of ASCII renderer core
- Establish ASCII renderer as pure rendering library
- Enable future fastn CLI integration via fastn-spec-viewer crate
- Maintain unified CLI behavior across development and production builds
- Clean separation allows focused development on each component

Architecture benefits:
✅ fastn-ascii-renderer - Pure ASCII rendering engine
✅ fastn-spec-viewer - Dedicated specification browser
✅ Clean dependencies - ASCII renderer focused on core rendering
✅ Future fastn integration - Via separate crate, not mixed concerns

Binary naming strategy:
- Development: cargo run -p fastn-spec-viewer
- Production: fastn spec-viewer (via CLI integration)
- Unified behavior across both execution methods
