+++ v0.5/Cargo.toml
@@ -3,6 +3,7 @@ members = [
     "fastn",
     "fastn-account",
     "fastn-ascii-renderer",
+    "fastn-spec-viewer",
     "fastn-automerge",
     "fastn-automerge-derive",
     "fastn-compiler",
