+++ v0.5/fastn-ascii-renderer/src/lib.rs
@@ -9,13 +9,6 @@
 // Placeholder usage for dependencies (will be used in later phases)
 use ansi_term as _;
 use unicode_width as _;
-use clap as _;
-use crossterm as _;  
-use notify as _;
-use once_cell as _;
-use ratatui as _;
-use syntect as _;
-use walkdir as _;
 
 mod canvas;
 mod layout;
@@ -26,8 +19,7 @@ mod ftd_types;
 mod css_mapper;
 mod ansi_canvas;
 
-#[cfg(feature = "spec-viewer")]
-pub mod spec_viewer;
+// spec_viewer module moved to separate fastn-spec-viewer crate
 
 pub use canvas::{Canvas, Position, Rect};
 pub use layout::{AsciiLayout, LayoutConstraints, ComponentLayout};
