--- v0.5/fastn-ascii-renderer/src/spec_viewer/mod.rs
@@ -1,3 +1,0 @@
-mod app;
-
-pub use app::SpecViewerApp;
